external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-webgl-identifier.vert-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: 0:2: 'webgl_' : reserved built-in name 

vertex shader that uses webgl identifier should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: vertex shader that uses webgl identifier should fail
test vertex shader
default fragment shaderPASS vertex shader that uses webgl identifier should fail
PASS successfullyParsed is true

TEST COMPLETE
